Refactor decorators' location
