+++ decorators/base_decorator.rb
